Support passing a custom `fetch` function
This commit adds support for passing a custom `fetch` function to the web client.

This is useful if you are not using global `fetch`, but rather need to use a different implementation or wrapper, for example for because you want to to use a custom HTTP agent/client to be able to configure mTLS or h1/h2 support settings.